+++ crates/foundry-ui/Cargo.toml
@@ -7,6 +7,7 @@ edition.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 anyhow.workspace = true
+starknet-types-core.workspace = true
 
 [dev-dependencies]
