+++ crates/foundry-ui/src/lib.rs
@@ -1,48 +1,9 @@
+use formats::{NumbersFormat, OutputFormat};
 pub use message::*;
 
-mod message;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum OutputFormat {
-    #[default]
-    Human,
-    Json,
-}
-
-impl OutputFormat {
-    #[must_use]
-    pub fn from_flag(json: bool) -> Self {
-        if json {
-            OutputFormat::Json
-        } else {
-            OutputFormat::Human
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum NumbersFormat {
-    Default,
-    Decimal,
-    Hex,
-}
-
-impl NumbersFormat {
-    #[must_use]
-    pub fn from_flags(hex_format: bool, dec_format: bool) -> Self {
-        assert!(
-            !(hex_format && dec_format),
-            "Exclusivity should be validated by clap"
-        );
-        if hex_format {
-            NumbersFormat::Hex
-        } else if dec_format {
-            NumbersFormat::Decimal
-        } else {
-            NumbersFormat::Default
-        }
-    }
-}
+pub mod formats;
+pub mod message;
+pub mod output_value;
 
 /// An abstraction around console output which stores preferences for output format (human vs JSON),
 /// colour, etc.
