+++ crates/sncast/src/starknet_commands/account/list.rs
@@ -3,8 +3,8 @@ use camino::Utf8PathBuf;
 use clap::Args;
 use conversions::string::IntoDecStr;
 use conversions::string::IntoHexStr;
-use foundry_ui::NumbersFormat;
-use foundry_ui::OutputFormat;
+use foundry_ui::formats::NumbersFormat;
+use foundry_ui::formats::OutputFormat;
 use itertools::Itertools;
 use serde::Deserialize;
 use serde::Serialize;
